+++ TSS.Java/src/tss/tpm/TPM2_PolicyCounterTimer_REQUEST.java
@@ -19,11 +19,11 @@ public class TPM2_PolicyCounterTimer_REQUEST extends TpmStructure
      * @param _offset the octet offset in the TPMS_TIME_INFO structure for the start of operand A 
      * @param _operation the comparison to make
      */
-    public TPM2_PolicyCounterTimer_REQUEST(TPM_HANDLE _policySession,byte[] _operandB,short _offset,TPM_EO _operation)
+    public TPM2_PolicyCounterTimer_REQUEST(TPM_HANDLE _policySession,byte[] _operandB,int _offset,TPM_EO _operation)
     {
         policySession = _policySession;
         operandB = _operandB;
-        offset = _offset;
+        offset = (short)_offset;
         operation = _operation;
     }
     /**
@@ -107,7 +107,7 @@ public class TPM2_PolicyCounterTimer_REQUEST extends TpmStructure
     {
         _p.add(d, "TPM_HANDLE", "policySession", policySession);
         _p.add(d, "byte", "operandB", operandB);
-        _p.add(d, "ushort", "offset", offset);
+        _p.add(d, "short", "offset", offset);
         _p.add(d, "TPM_EO", "operation", operation);
     };
